hash combine with size_t

+++ src/nox/netapps/fns/rules.cc
@@ -52,7 +52,7 @@ void EPoint::installed_pop() {
 
 uint64_t EPoint::generate_key(uint64_t sw_id, uint32_t port, uint16_t vlan,
 		uint32_t mpls) {
-	uint64_t seed = 0;
+	size_t seed = 0;
 	boost::hash_combine(seed, port);
 	boost::hash_combine(seed, sw_id);
 	boost::hash_combine(seed, vlan);
